Move an indication parameter to IconButtonStateStyle
