+++ component/compose-stylesheet-button/src/commonMain/kotlin/com.moriatsushi.compose.stylesheet.button/IconButtonStateStyleBuilder.kt
@@ -4,6 +4,7 @@ import androidx.compose.ui.graphics.Color
 import com.moriatsushi.compose.stylesheet.StyleSheetBuilderMarker
 import com.moriatsushi.compose.stylesheet.component.ComponentCommonStyleBuilder
 import com.moriatsushi.compose.stylesheet.component.ComponentCommonStyleHelper
+import com.moriatsushi.compose.stylesheet.indication.IndicationSetter
 import com.moriatsushi.compose.stylesheet.token.TokenSetter
 
 /**
@@ -18,17 +19,31 @@ open class IconButtonStateStyleBuilder internal constructor(
      */
     val color = TokenSetter<Color>()
 
+    /**
+     * An indication representing visual effects that occur when certain interactions happen, such
+     * as pressing.
+     *
+     * Example:
+     * ```
+     * indication += SampleIndication()
+     * indication += { rememberRipple() }
+     * ```
+     */
+    val indication: IndicationSetter = IndicationSetter()
+
     operator fun invoke(builder: IconButtonStateStyleBuilder.() -> Unit) {
         builder()
     }
 
     internal operator fun plusAssign(other: IconButtonStateStyle) {
         color += other.color
+        indication += other.indication
         this += other.commonStyle
     }
 
     internal fun buildStateStyle(): IconButtonStateStyle = IconButtonStateStyle(
         color = color.token,
+        indication = indication.value,
         commonStyle = commonStyleHelper.buildCommonStyle(),
     )
 }
